Restructure modules... again.
