+++ src/geometry/types.rs
@@ -1,0 +1,240 @@
+use std::ops::{ Add, AddAssign, Mul, Div };
+
+// Point /////////////////////////////////////////////////////////////////////
+//
+// Coordinates in 2D space.
+
+#[derive(PartialEq, Debug)]
+pub struct Point {
+    pub x: i32,
+    pub y: i32,
+}
+
+
+//impl PartialEq for Point {
+//    fn eq(&self, other: &'_ Point) -> bool {
+//        self.x == other.x && self.y == other.y
+//    }
+//}
+
+impl Point {
+    pub fn origin() -> Point {
+        Point { x: 0, y: 0 }
+    }
+
+    pub fn is_origin(&self) -> bool {
+        self == &Point::origin()
+    }
+
+    pub fn distance_to(&self, other: &Point) -> f64 {
+        let difference = Vector::difference(self, other);
+        difference.magnitude()
+    }
+}
+
+// Vector ////////////////////////////////////////////////////////////////////
+//
+// Change of coordinates in 2D space.
+
+#[derive(Debug)]
+pub struct Vector {
+    pub dx: f64,
+    pub dy: f64,
+}
+
+impl PartialEq for Vector {
+    fn eq(&self, other: &'_ Vector) -> bool {
+        let e = 0.0000001;
+        let x = (self.dx - other.dx).abs();
+        let y = (self.dy - other.dy).abs();
+        x < e && y < e
+    }
+}
+
+impl Add for Vector {
+    type Output = Vector;
+
+    fn add(self, rhs: Vector) -> Self::Output {
+        Vector {
+            dx: self.dx + rhs.dx,
+            dy: self.dy + rhs.dy,
+        }
+    }
+}
+
+impl AddAssign for Vector {
+    fn add_assign(&mut self, rhs: Vector) {
+        self.dx += rhs.dx;
+        self.dy += rhs.dy;
+    }
+}
+
+impl<'a> Mul<f64> for &'a Vector {
+    type Output = Vector;
+
+    fn mul(self, scalar: f64) -> Self::Output {
+        Vector {
+            dx: self.dx * scalar,
+            dy: self.dy * scalar,
+        }
+    }
+}
+
+impl<'a> Div<f64> for &'a Vector {
+    type Output = Vector;
+
+    fn div(self, scalar: f64) -> Self::Output {
+        Vector {
+            dx: self.dx / scalar,
+            dy: self.dy / scalar,
+        }
+    }
+}
+
+impl<'a> Mul for &'a Vector {
+    type Output = f64;
+
+    fn mul(self, rhs: &Vector) -> Self::Output {
+        self.dx * rhs.dx + self.dy * rhs.dy
+    }
+}
+
+impl Vector {
+    pub fn zero() -> Vector {
+        Vector { dx: 0.0, dy: 0.0 }
+    }
+
+    pub fn difference(lhs: &Point, rhs: &Point) -> Vector {
+        Vector {
+            dx: (lhs.x - rhs.x) as f64,
+            dy: (lhs.y - rhs.y) as f64,
+        }
+    }
+
+    pub fn magnitude(&self) -> f64 {
+        ((self.dx * self.dx + self.dy * self.dy) as f64).sqrt()
+    }
+
+    pub fn normalized(&self) -> Option<Vector> {
+        if self == &Vector::zero() { return None }
+        let magnitude = self.magnitude();
+
+        Some(Vector {
+            dx: self.dx / magnitude,
+            dy: self.dy / magnitude
+        })
+    }
+}
+
+// Tests /////////////////////////////////////////////////////////////////////
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn point_distance_from_origin() {
+        // given
+        let p1 = Point { x: 0, y: 0 };
+        let p2 = Point { x: 5, y: 0 };
+
+        // when
+        let result = p1.distance_to(&p2);
+
+        // then
+        assert_eq!(result, 5.0);
+    }
+
+    #[test]
+    fn point_distance_to_origin() {
+        // given
+        let p1 = Point { x: 0, y: 0 };
+        let p2 = Point { x: 0, y: -5 };
+
+        // when
+        let result = p2.distance_to(&p1);
+
+        // then
+        assert_eq!(result, 5.0);
+    }
+
+    // Vector ////////////////////////////////////////////////////////////////
+
+    #[test]
+    fn vector_add_assigns() {
+        // given
+        let mut sut = Vector { dx: 3.0, dy: 4.0 };
+
+        // when
+        sut += Vector { dx: 9.5, dy: -3.5 };
+
+        // then
+        assert_eq!(sut, Vector { dx: 12.5, dy: 0.5 });
+    }
+
+    #[test]
+    fn vector_scalar_multiplies() {
+        // given
+        let sut = Vector { dx: 3.0, dy: 4.0 };
+
+        // when
+        let result = &sut * 3.0;
+
+        // then
+        assert_eq!(result, Vector { dx: 9.0, dy: 12.0 })
+    }
+
+    #[test]
+    fn vector_scalar_divides() {
+        // given
+        let sut = Vector { dx: 3.0, dy: 12.0 };
+
+        // when
+        let result = &sut / 3.0;
+
+        // then
+        assert_eq!(result, Vector { dx: 1.0, dy: 4.0 });
+    }
+
+    #[test]
+    fn vector_inner_product() {
+        // given
+        let a = Vector { dx: 3.4, dy: -4.9 };
+        let b = Vector { dx: 10.0, dy: 6.3 };
+
+        // when
+        let result = &a * &b;
+
+        // then
+        assert!((result - 3.13).abs() < 0.0000001);
+    }
+
+    #[test]
+    fn vector_magnitude() {
+        // given, when, then
+        assert_eq!(Vector { dx: 3.0, dy: 4.0 }.magnitude(), 5.0)
+    }
+
+    #[test]
+    fn vector_normalize() {
+        // given
+        let sut = Vector { dx: 3.3, dy: 5.2 };
+
+        // when
+        match sut.normalized() {
+            None => {
+                panic!()
+            },
+            Some(result) => {
+                // then
+                assert!((result.magnitude() - 1.0).abs() < 0.0000001);
+            }
+        };
+    }
+
+    #[test]
+    fn vector_does_not_normalize_if_zero() {
+        // given, when, then
+        assert_eq!(Vector::zero().normalized(), None)
+    }
+}
