+++ src/geometry/util.rs
@@ -1,7 +1,41 @@
-use physics::body::Body;
-use geometry::{ point::Point, vector::Vector, transformation::Transformation };
+use physics::types::Body;
+use geometry::types::{ Point, Vector};
 use rand::prelude::*;
 use std::f64::consts::PI;
+use std::ops::Mul;
+
+// Transformation ////////////////////////////////////////////////////////////
+//
+// A 2D transformation matrix represented as a pair of transformed basis
+// vectors.
+
+// TODO: make this a data tuple
+pub struct Transformation {
+    pub a: Vector,
+    pub b: Vector,
+}
+
+impl<'a> Mul<Vector> for &'a Transformation {
+    type Output = Vector;
+
+    fn mul(self, rhs: Vector) -> Self::Output {
+        &self.a * rhs.dx + &self.b * rhs.dy
+    }
+}
+
+impl Transformation {
+    pub fn rotation(radians: f64) -> Transformation {
+        let (sin, cos) = radians.sin_cos();
+        Transformation {
+            a: Vector { dx: cos, dy: sin },
+            b: Vector { dx: -sin, dy: cos },
+        }
+    }
+}
+
+// Distributor ///////////////////////////////////////////////////////////////
+//
+// // TODO: description
 
 pub struct Distributor {
     pub num_bodies: u32,
@@ -11,10 +45,7 @@ pub struct Distributor {
 }
 
 impl Distributor {
-    /**
-     *  Returns a distribution of bodies.
-     */
-    pub fn get(&self) -> Vec<Body> {
+    pub fn distribution(&self) -> Vec<Body> {
         let mut result: Vec<Body> = vec![];
         let mut angle_rand = thread_rng();
         let mut dist_rand = thread_rng();
@@ -39,5 +70,41 @@ impl Distributor {
         }
 
         result
+    }
+}
+
+// Tests /////////////////////////////////////////////////////////////////////
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use geometry::types::Vector;
+    use std::f64::consts::FRAC_PI_2;
+
+    #[test]
+    fn it_transforms_a_vector() {
+        // given
+        let sut = Transformation {
+            a: Vector { dx: 2.0, dy: 0.0 },
+            b: Vector { dx: 0.0, dy: 2.0 },
+        };
+
+        // when
+        let result = &sut * Vector { dx: 4.0, dy: -2.5 };
+
+        // then
+        assert_eq!(result, Vector { dx: 8.0, dy: -5.0 });
+    }
+
+    #[test]
+    fn it_rotates_a_vector() {
+        // given
+        let sut = Transformation::rotation(FRAC_PI_2);
+
+        // when
+        let result = &sut * Vector { dx: 1.0, dy: 0.0 };
+
+        // then
+        assert_eq!(result, Vector { dx: 0.0, dy: 1.0 });
     }
 }
