--- src/physics/body.rs
@@ -1,38 +1,0 @@
-use std::cmp::Eq;
-use geometry::{ point::Point, vector::Vector };
-
-#[derive(Debug)]
-pub struct Body {
-    pub id: u32,
-    pub mass: f64,          // TODO: need to ensure it is positive & nonzero
-    pub position: Point,
-    pub velocity: Vector,
-}
-
-impl Eq for Body {}
-
-impl PartialEq for Body {
-    /**
-     *  Bodies are compared referentially.
-     */
-    fn eq(&self, other: &'_ Body) -> bool {
-        self as *const _ == other as *const _
-    }
-}
-
-impl Drop for Body {
-    fn drop(&mut self) {
-        println!("Body #{} has been deallocated.", self.id);
-    }
-}
-
-impl Body {
-    /**
-     *  Apply the given force to update the velocity vector.
-     */
-    pub fn apply_force(&mut self, force: &Vector) {
-        self.velocity += force / self.mass;
-        self.position.x += self.velocity.dx.round() as i32;
-        self.position.y += self.velocity.dy.round() as i32;
-    }
-}
