+++ src/physics/mod.rs
@@ -1,2 +1,1 @@
-pub mod body;
-pub mod field;
+pub mod types;
