+++ src/physics/types.rs
@@ -1,6 +1,41 @@
-use physics::body::Body;
-use geometry::vector::Vector;
+use geometry::types::{ Point, Vector };
+use std::cmp::Eq;
 use std::collections::HashMap;
+
+// Body //////////////////////////////////////////////////////////////////////
+//
+// // TODO: description
+
+#[derive(Debug)]
+pub struct Body {
+    pub id: u32,
+    pub mass: f64,          // TODO: need to ensure it is positive & nonzero
+    pub position: Point,
+    pub velocity: Vector,
+}
+
+impl Eq for Body {
+
+}
+
+impl PartialEq for Body {
+    fn eq(&self, other: &'_ Body) -> bool {
+        // Bodies are compared referentially.
+        self as *const _ == other as *const _
+    }
+}
+
+impl Body {
+    pub fn apply_force(&mut self, force: &Vector) {
+        self.velocity += force / self.mass;
+        self.position.x += self.velocity.dx.round() as i32;
+        self.position.y += self.velocity.dy.round() as i32;
+    }
+}
+
+// Field /////////////////////////////////////////////////////////////////////
+//
+//
 
 pub struct Field {
     pub g: f64,
@@ -10,14 +45,7 @@ pub struct Field {
     pub bodies: Vec<Body>,
 }
 
-impl Drop for Field {
-    fn drop(&mut self) {
-        println!("A field has been deallocated.");
-    }
-}
-
 impl Field {
-
     // TODO: Needs testing
     /**
      *  Update the state of the field by applying force on each of the bodies
@@ -103,5 +131,55 @@ impl Field {
         };
 
         &direction * force
+    }
+}
+
+// Tests /////////////////////////////////////////////////////////////////////
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use geometry::types::{ Point, Vector };
+
+    #[test]
+    fn body_has_referential_equivalence() {
+        // given
+        let b1 = Body {
+            id: 0,
+            mass: 1.0,
+            position: Point { x: 1, y: 2 },
+            velocity: Vector::zero(),
+        };
+
+        let b2 = Body {
+            id: 0,
+            mass: 1.0,
+            position: Point { x: 1, y: 2 },
+            velocity: Vector::zero(),
+        };
+
+        // then
+        assert_eq!(b1, b1);
+        assert_ne!(b1, b2);
+    }
+
+    #[test]
+    fn body_applies_force() {
+        // given
+        let mut sut = Body {
+            id: 0,
+            mass: 2.0,
+            position: Point { x: 1, y: 2 },
+            velocity: Vector { dx: -2.0, dy: 5.0 },
+        };
+
+        let force = Vector { dx: 2.6, dy: -3.2 };
+
+        // when
+        sut.apply_force(&force);
+
+        // then
+        assert_eq!(sut.velocity, Vector { dx: -0.7, dy: 3.4 });
+        assert_eq!(sut.position, Point { x: 0, y: 5 });
     }
 }
