--- src/tests/mod.rs
@@ -1,4 +1,0 @@
-mod vector_tests;
-mod point_tests;
-mod body_tests;
-mod transformation_tests;
